Attempting to avoid the awful UI for 360Fly Director
It forces me to process one video at a time, and when it's done, I can click on the next one and tell it to
process. This is unacceptable. It's a waste of my time. I should be able to just tell it to run and go to bed, then wake
up in the morning to a fresh batch of equirectangular files.

I believe this code will do just that, at least for a single directory. The downside is that this does not add in the
"blind spot", so I have to edit that together in Davinci. I'm hoping I can just combine all the clips into a single
compound clip and do the resize all at once, layering in a still from some footage previously exported by Director.