A second step is required for davinci to read it
for whatever reason, the output of the first command is not quite good enough for davinci. I did some googling and came
up with this, which it does import. I did try making a combined version of this command, so it would remap and convert
in 1 step, but the output was black and white. That doesn't make sense, but for some reason making it two steps Just
Works.

Also, get rid of the debugging code, i.e. "echo" - derp.