-c:v mpeg4 caused ffmpeg to complain the headers were damaged
and it would refuse to work at all. I bashed my head on the wall for so friggin long trying to get it to run like that,
forgetting I had made that change in an attempt to get Davinci to recognize the output off the hop. Holy frig.

Oh, the second change about -b:v250000k in the blindspotting part is because I was noticing the bitrate dropped off
during this step huge. I need to actually reduce that because the output filesize is larger than the input, but I need
to research how to get that right.