this bitrate seems closer to what the camera uses
I watched some raw footage in VLC and paid attention to the bitrate statistics, then rounded up. The output is only a
little bit bigger than the raw file from the camera, whereas previously it was growing by a lot. When my first 4tb
hard-drive got close to full I realized that tolerating this was costing me money!

I think the reason for the file size growth is just that when I inject a blindspot, ffmpeg is having to actually create
pixels and then encode them at the specified bitrate, whereas when it processes video, it just takes as much as it can
get up to the specified bitrate, so 69000k from the camera would get encoded as 69000k, but the blindspot area was being
encoded at 250000k thus bloating the output.

I've also applied the bitrate argument to the remap step as the output was actually decreasing in filesize at this
stage, which I can only presume means the quality is suffering. I'm not really an expert in video encodings.